relocated Service.java to impl module (along with tests).
