--- api/src/test/resources/META-INF/services/io.jsonwebtoken.StubService
@@ -1,1 +1,0 @@
-io.jsonwebtoken.DefaultStubService
