+++ impl/src/main/java/io/jsonwebtoken/impl/lang/LegacyServices.java
@@ -1,8 +1,6 @@
 package io.jsonwebtoken.impl.lang;
 
 import io.jsonwebtoken.lang.Classes;
-import io.jsonwebtoken.lang.Services;
-import io.jsonwebtoken.lang.UnavailableImplementationException;
 import io.jsonwebtoken.lang.UnknownClassException;
 
 /**
