+++ impl/src/test/resources/META-INF/services/io.jsonwebtoken.StubService
@@ -1,0 +1,1 @@
+io.jsonwebtoken.impl.DefaultStubService
